vision/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
Brlapi==0.6.6
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
command-not-found==0.3
companies-house==0.1.2
cryptography==2.1.4
cssselect==1.0.3
cupshelpers==1.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.6
django-crispy-forms==1.7.2
entrypoints==0.3
factory-boy==2.12.0
Faker==1.0.7
falcon==2.0.0
Flask==1.1.1
freeze==1.0.10
httplib2==0.9.2
idna==2.8
importlib-metadata==0.17
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.3.3
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.0.0
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
notebook==5.7.8
numpy==1.16.4
oauth==1.0.1
olefile==0.45.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.1
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plotly==3.10.0
pluggy==0.12.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.0.0
ptyprocess==0.6.0
py==1.8.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.0
pyRFC3339==1.0
pyrsistent==0.15.2
pytest==4.6.1
pytest-django==3.4.8
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.1
qtconsole==4.5.1
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
retrying==1.3.3
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.13.2
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.2
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
virtualenv==16.7.5
w3lib==1.20.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
xkit==0.0.0
zipp==0.5.1
zope.interface==4.3.2
